Fix the webcam device problem
Before, when you ran this program with webcam, it will crush with the error 'Nonetype object has no attribute 'shape'.
It happened because the type of 'video_path' is string, but cv2.VideoCapture(0) need it to be 'int' type.
So I made a little adjustment at line 37 ~ 40 and it work successfully.